Python/requirements.txt
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pycodestyle==2.6.0
python-Levenshtein-wheels==0.13.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
